+++ prp/models/qc.py
@@ -23,6 +23,7 @@ class QcSoftware(Enum):
     CHEWBBACA = TypingSoftware.CHEWBBACA.value
     GAMBITCORE = "gambitcore"
     NANOPLOT = "nanoplot"
+    SAMTOOLS = "samtools"
 
 
 class QuastQcResult(BaseModel):
@@ -88,6 +89,23 @@ class NanoPlotQcResult(BaseModel):
     total_bases: float
 
 
+class ContigCoverage(BaseModel):
+    """Coverage information for a single contig."""
+    rname: str
+    startpos: int
+    endpos: int
+    numreads: int
+    covbases: int
+    coverage: float
+    meandepth: float
+    meanbaseq: float
+    meanmapq: float
+
+class SamtoolsCoverageQcResult(BaseModel):
+    """SAMtools coverage QC result model."""
+    contigs: list[ContigCoverage]
+
+
 class QcMethodIndex(RWModel):
     """QC results container.
 
@@ -97,7 +115,7 @@ class QcMethodIndex(RWModel):
 
     software: QcSoftware
     version: str | None = None
-    result: QuastQcResult | PostAlignQcResult | GenomeCompleteness | GambitcoreQcResult | NanoPlotQcResult
+    result: QuastQcResult | PostAlignQcResult | GenomeCompleteness | GambitcoreQcResult | NanoPlotQcResult | SamtoolsCoverageQcResult
 
 
 class CdmQcMethodIndex(QcMethodIndex):