+++ tests/fixtures/saureus/__init__.py
@@ -109,3 +109,8 @@ def saureus_sccmec_path(data_path: Path) -> str:
 def saureus_nanoplot_path(data_path: Path) -> str:
     """Get path for saureus nanoplot file"""
     return str(data_path.joinpath("saureus", "nanoplot.txt"))
+
+@pytest.fixture()
+def saureus_samtools_coverage_path(data_path: Path) -> str:
+    """Get path for saureus samtools coverage file"""
+    return str(data_path.joinpath("saureus", "samcoverage.txt"))